Fixed check_averaging circular import and sorted import ordering
