+++ gprMax/cmds_geometry/box.py
@@ -23,9 +23,9 @@ import numpy as np
 import gprMax.config as config
 
 from ..cython.geometry_primitives import build_box
-from ..hash_cmds_geometry import check_averaging
 from ..materials import Material
-from .cmds_geometry import UserObjectGeometry, rotate_2point_object
+from .cmds_geometry import (UserObjectGeometry, check_averaging,
+                            rotate_2point_object)
 
 logger = logging.getLogger(__name__)
