+++ gprMax/cmds_geometry/fractal_box.py
@@ -21,11 +21,13 @@ import logging
 import numpy as np
 
 import gprMax.config as config
-from gprMax.cmds_geometry.cmds_geometry import UserObjectGeometry, rotate_2point_object
+from gprMax.cmds_geometry.cmds_geometry import (UserObjectGeometry,
+                                                rotate_2point_object)
 from gprMax.fractals import FractalVolume
 from gprMax.materials import ListMaterial
 
-from ..cython.geometry_primitives import build_voxels_from_array, build_voxels_from_array_mask
+from ..cython.geometry_primitives import (build_voxels_from_array,
+                                          build_voxels_from_array_mask)
 
 logger = logging.getLogger(__name__)
