+++ gprMax/cmds_geometry/plate.py
@@ -20,7 +20,8 @@ import logging
 
 import numpy as np
 
-from ..cython.geometry_primitives import build_face_xy, build_face_xz, build_face_yz
+from ..cython.geometry_primitives import (build_face_xy, build_face_xz,
+                                          build_face_yz)
 from .cmds_geometry import UserObjectGeometry, rotate_2point_object
 
 logger = logging.getLogger(__name__)
