+++ gprMax/cmds_geometry/triangle.py
@@ -21,9 +21,8 @@ import logging
 import numpy as np
 
 from ..cython.geometry_primitives import build_triangle
-from ..hash_cmds_geometry import check_averaging
 from ..materials import Material
-from .cmds_geometry import UserObjectGeometry, rotate_point
+from .cmds_geometry import UserObjectGeometry, check_averaging, rotate_point
 
 logger = logging.getLogger(__name__)
